// src/modules/common/enums/s3-file-upload-features.enum.ts
export enum FileUploadFeatureType {
  POSTS = 'posts',
  USER_AVATAR = 'user/avatar',
  COMPANY_LOGO = 'company/avatar',
  CHANNEL_BANNER = 'channel/banner',
  CHANNEL_ICON = 'channel/icon',
}

export enum FileUploadRequestTypes {
  POSTS = 'posts',
  USER_AVATAR = 'userAvatar',
  COMPANY_LOGO = 'companyLogo',
  CHANNEL_BANNER = 'channelBanner',
  CHANNEL_ICON = 'channelIcon',
}